+++ bindings/builtin/symbolizer.go
@@ -1,5 +1,3 @@
-// +build cgosymbolizer
-
 package builtin
 
 // extern void cgoTraceback(void*);
@@ -7,11 +5,14 @@ package builtin
 // extern void cgoSignalsInit();
 import "C"
 import (
+	"os"
 	"runtime"
 	"unsafe"
 )
 
 func init() {
-	C.cgoSignalsInit()
-	runtime.SetCgoTraceback(0, unsafe.Pointer(C.cgoTraceback), nil, unsafe.Pointer(C.cgoSymbolizer))
+	if os.Getenv("REINDEXER_CGOBACKTRACE") != "" {
+		C.cgoSignalsInit()
+		runtime.SetCgoTraceback(0, unsafe.Pointer(C.cgoTraceback), nil, unsafe.Pointer(C.cgoSymbolizer))
+	}
 }